Fix BibTeX error parsing case
Given an error like

I couldn't open auxiliary file biblio.aux
---line 3 of file latex.out/main.aux

the regexp this previously used would consume the first '-' as an
error message and thus ignore the previous line.  Add lazyness to the
regexp so that the message on the second line is the empty string and
we fall back to the previous line.