Report position on line in error stacks
Previously, latexrun only reported the "before" part of each error
stack line.  Often this was enough, but sometimes the important part
is in the "after" part and this was very confusing (and often required
looking in the log to find the "after" part).  Fix this by reading and
including the "after" part.  Use a ^ marker to indicate where on the
line the error was, which is the more modern convention for this.