Clarify hidden warning message using -Wall
Previously, this just said to "use -W" to show hidden warnings.  This
was intended to mean that the user should use -W with the appropriate
arguments to show the warnings, but taken at face value, it suggested
illegal usage.  Fix this by suggesting -Wall.  This is legal usage and
the user can dig deeper into -W if they wish.

Closes #7