Force re-run on missing \include
Previously, missing \includes did not affect dependencies at all, so
if the file was later created, latexrun could still consider the task
stable and not re-run it.

It's quite hard to model what goes in to LaTeX deciding that a
\include file is missing and not super important to get right, so fix
this by depending on the conservative "unknown" input.  This way,
latexrun will always re-run the latex task at least once the next time
it's invoked, giving latex a change to pick up newly created files.

While there's no portable way to address the missing input file
problem in general, this at least addresses the most visible and
likely manifestation.

Closes #8